TypeScript migration, update rules, add getAuthObject()
Add firestore rules to require timetypeName along with timeType and to require that jobDescription and client are present when job is defined.

update eslint dependency and fix the script command

linting

migrate timesheets.js to timesheets.ts

create getAuthObject() for usage in function. possibly use it to later replace hasPermission()